update: redis as package
